# docs/swagger.yaml
basePath: /
definitions:
  github_com_waanvieira_api-users_internal_dto.CreateProductInput:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  github_com_waanvieira_api-users_internal_dto.CreateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  github_com_waanvieira_api-users_internal_dto_users.GetJWTInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_waanvieira_api-users_internal_dto_users.GetJWTOutput:
    properties:
      access_token:
        type: string
    type: object
  github_com_waanvieira_api-users_internal_entity.Product:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  internal_infra_webserver_handlers.Error:
    properties:
      message:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    name: Wanderson Vieira
  description: Product API with auhtentication
  license:
    name: W
  termsOfService: http://swagger.io/terms/
  title: Go Expert API Example
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get all products
      parameters:
      - description: page number
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_waanvieira_api-users_internal_entity.Product'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create products
      parameters:
      - description: product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_waanvieira_api-users_internal_dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Create product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product
      parameters:
      - description: product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get a product
      parameters:
      - description: product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_waanvieira_api-users_internal_entity.Product'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update a product
      parameters:
      - description: product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_waanvieira_api-users_internal_dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Update a product
      tags:
      - products
  /users:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_waanvieira_api-users_internal_dto.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      summary: Create user
      tags:
      - users
  /users/generate_token:
    post:
      consumes:
      - application/json
      description: Get a user JWT
      parameters:
      - description: user credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_waanvieira_api-users_internal_dto_users.GetJWTInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_waanvieira_api-users_internal_dto_users.GetJWTOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_infra_webserver_handlers.Error'
      summary: Get a user JWT
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"